+++ src/futhark.hs
@@ -281,12 +281,6 @@ commandLineOptions =
   , Option [] ["defunctionalise"]
     (NoArg $ Right $ \opts -> opts { futharkPipeline = Defunctionalise })
     "Defunctionalise the program."
-  , Option "I" ["include"]
-    (ReqArg (\path -> Right $ changeFutharkConfig $ \opts ->
-                opts { futharkImportPaths =
-                         futharkImportPaths opts `mappend` importPath path })
-    "DIR")
-    "Add directory to search path."
   , typedPassOption soacsProg Kernels firstOrderTransform "f"
   , soacsPassOption fuseSOACs "o"
   , soacsPassOption inlineAndRemoveDeadFunctions []
@@ -333,7 +327,7 @@ main = mainWithOptions newConfig commandLineOptions compile
           case futharkPipeline config of
             TypeCheck -> do
               -- No pipeline; just read the program and type check
-              (warnings, _, _) <- readProgram preludeBasis mempty file
+              (warnings, _, _) <- readProgram preludeBasis file
               liftIO $ hPutStr stderr $ show warnings
             PrettyPrint -> liftIO $ do
               maybe_prog <- parseFuthark file <$> T.readFile file
@@ -341,16 +335,16 @@ main = mainWithOptions newConfig commandLineOptions compile
                 Left err  -> fail $ show err
                 Right prog-> putStrLn $ pretty prog
             Defunctorise -> do
-              (_, imports, src) <- readProgram preludeBasis mempty file
+              (_, imports, src) <- readProgram preludeBasis file
               liftIO $ mapM_ (putStrLn . pretty) $
                 evalState (Defunctorise.transformProg imports) src
             Monomorphise -> do
-              (_, imports, src) <- readProgram preludeBasis mempty file
+              (_, imports, src) <- readProgram preludeBasis file
               liftIO $ mapM_ (putStrLn . pretty) $ flip evalState src $
                 Defunctorise.transformProg imports
                 >>= Monomorphise.transformProg
             Defunctionalise -> do
-              (_, imports, src) <- readProgram preludeBasis mempty file
+              (_, imports, src) <- readProgram preludeBasis file
               liftIO $ mapM_ (putStrLn . pretty) $ flip evalState src $
                 Defunctorise.transformProg imports
                 >>= Monomorphise.transformProg